Refactor turret subsystem: remove controller dependency and improve shot compensation logic
