+++ src/main/java/frc/robot/subsystems/turret/TurretIOSensorInputs.java
@@ -13,6 +13,7 @@ import edu.wpi.first.wpilibj.DutyCycleEncoder;
 import frc.lib.subsystem.IORefresher;
 import frc.robot.CanID;
 import frc.robot.RobotContainer;
+import frc.robot.subsystems.turret.calc.TurretMath;
 import frc.robot.subsystems.vision.photon.Camera;
 
 import org.littletonrobotics.junction.Logger;
